requirements.txt
openai==1.3.7
beautifulsoup4==4.12.2
requests==2.31.0
chromadb==0.4.18
langchain==0.0.345
langchain-openai==0.0.2
python-dotenv==1.0.0
streamlit==1.28.1
tiktoken==0.5.1
lxml==4.9.3
scrapy==2.11.0
selenium==4.15.2
webdriver-manager==4.0.1
numpy==1.24.3
pandas==2.0.3